Revert "Making master the main branch"
