+++ Databases_Engine/MongoLink.cs
@@ -3,187 +3,95 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using BHoM.Base;
 using BHoM.Global;
 using MongoDB.Bson;
 using MongoDB.Driver;
-using BHoM.Databases;
-using System.Text.RegularExpressions;
-using System.IO;
 
 namespace Mongo_Adapter
 {
-    public class MongoLink : IDatabaseCollectionAdapter
+    public class MongoLink
     {
-        private MongoClient m_Client;
-        private IMongoCollection<BsonDocument> m_Collection;
-        private IMongoCollection<BsonDocument> m_History;
+        private IMongoCollection<BsonDocument> collection;
+        private IMongoCollection<BsonDocument> depCollection;
 
         public MongoLink(string serverLink = "mongodb://localhost:27017", string databaseName = "project", string collectionName = "bhomObjects")
         {
-            if (!serverLink.StartsWith("mongodb://"))
-                serverLink = "mongodb://" + serverLink + ":27017";
-
-            m_Client = new MongoClient(serverLink);
-            IMongoDatabase database = m_Client.GetDatabase(databaseName);
-            m_Collection = database.GetCollection<BsonDocument>(collectionName);
-
-            HistorySize = 20;
-
-            IMongoDatabase hist_Database = m_Client.GetDatabase(databaseName + "_History");
-            m_History = hist_Database.GetCollection<BsonDocument>(collectionName);
+            var mongo = new MongoClient(serverLink);
+            IMongoDatabase database = mongo.GetDatabase(databaseName);
+            collection = database.GetCollection<BsonDocument>(collectionName);
+            depCollection = database.GetCollection<BsonDocument>(collectionName + "__dep");
         }
 
-        /*******************************************/
-
-        public string ServerName
+        public void SaveObjects(IEnumerable<BHoMObject> objects, string key = "")
         {
-            get
-            {
-                MongoServerAddress server = m_Collection.Database.Client.Settings.Server;
-                return "mongodb://" + server.ToString();
-            }
-        }
-
-        /*******************************************/
-
-        public string DatabaseName
-        {
-            get { return m_Collection.Database.DatabaseNamespace.DatabaseName;  }
-        }
-
-        /*******************************************/
-
-        public string CollectionName
-        {
-            get { return m_Collection.CollectionNamespace.CollectionName;  }
-        }
-
-        /*******************************************/
-
-        public int HistorySize { get; set; }
-
-        /*******************************************/
-
-        public bool Push(IEnumerable<object> objects, string key, List<string> tags = null)
-        {
-            if (m_Client.Cluster.Description.State == MongoDB.Driver.Core.Clusters.ClusterState.Disconnected)
-                return false;
-
             // Create the bulk query for the object to replace/insert
-            DateTime timestamp = DateTime.Now; 
             List<WriteModel<BsonDocument>> bulk = new List<WriteModel<BsonDocument>>();
-            WriteModel<BsonDocument> deletePrevious = new DeleteManyModel<BsonDocument>(Builders<BsonDocument>.Filter.Eq("__Key__", key));
-            bulk.Add(deletePrevious);
-            foreach (object obj in objects)
-                bulk.Add(new InsertOneModel<BsonDocument>(ToBson(obj, key, timestamp)));
+            bulk.Add(new DeleteManyModel<BsonDocument>(Builders<BsonDocument>.Filter.Eq("Key", key)));
+            foreach (BHoMObject obj in objects)
+                bulk.Add(new InsertOneModel<BsonDocument>(ToBson(obj, key)));
 
             // Send that query
             BulkWriteOptions bulkOptions = new BulkWriteOptions();
             bulkOptions.IsOrdered = true;
-            m_Collection.BulkWrite(bulk, bulkOptions);
+            collection.BulkWrite(bulk, bulkOptions);
 
-            // Push in the history database as well
-            bulk.Remove(deletePrevious);
-            List<object> times = Query(new List<string> { "{$group: {_id: \"$__Time__\"}}", "{$sort: {_id: -1}}" });
-            if (times.Count > HistorySize)
+            // Get all the dependencies
+            Dictionary<Guid, BHoMObject> dependencies = new Dictionary<Guid, BHoMObject>();
+            foreach (BHoMObject bhomObject in objects)
             {
-                bulk.Insert(0, new DeleteManyModel<BsonDocument>(Builders<BsonDocument>.Filter.Lte("__Time__", times[HistorySize])));
+                foreach (KeyValuePair<Guid, BHoMObject> kvp in bhomObject.GetDeepDependencies())
+                {
+                    if (!dependencies.ContainsKey(kvp.Key))
+                        dependencies[kvp.Key] = kvp.Value;
+                }
             }
 
-            m_History.BulkWrite(bulk, bulkOptions);
-            return true;
+            // Create the bulk query for the dependencies to replace/insert
+            List<WriteModel<BsonDocument>> depBulk = new List<WriteModel<BsonDocument>>();
+            depBulk.Add(new DeleteManyModel<BsonDocument>(Builders<BsonDocument>.Filter.Eq("Key", key)));
+            foreach (BHoMObject obj in dependencies.Values)
+                depBulk.Add(new InsertOneModel<BsonDocument>(ToBson(obj, key)));
+
+            // Send that query
+            depCollection.BulkWrite(depBulk, bulkOptions);
         }
 
-        /*******************************************/
-
-        public bool Delete(string filterString = "{}")
+        public void DeleteObjects(string filterString)
         {
-            if (m_Client.Cluster.Description.State == MongoDB.Driver.Core.Clusters.ClusterState.Disconnected)
-                return false;
-
             FilterDefinition<BsonDocument> filter = filterString;
-            m_Collection.DeleteMany(filter);
-            return true;
+            collection.DeleteMany(filter);
         }
 
-        /*******************************************/
-
-        public List<object> Pull(string filterString = "{}", bool keepAsString = false)
+        public void Clear()
         {
-            if (m_Client.Cluster.Description.State == MongoDB.Driver.Core.Clusters.ClusterState.Disconnected)
-                return new List<object>();
+            collection.DeleteMany(new BsonDocument());
+        }
 
+        public IEnumerable<BHoMObject> GetObjects(string filterString)
+        {
+            // Add the queried objects to a temp project
+            Project tempProject = new Project();
             FilterDefinition<BsonDocument> filter = filterString;
-            List<BsonDocument> result = m_Collection.Find(filter).ToList();
-            if (keepAsString)
-                return result.Select(x => x.ToString()).ToList<object>();
-            else
-                return result.Select(x => FromBson(x)).ToList<object>();
+            var result = collection.Find(filter);
+            var ret = result.ToList().Select(x => BHoMObject.FromJSON(x.ToString(), tempProject));
+            foreach (BHoMObject obj in ret)
+                tempProject.AddObject(obj);
+
+            // Sort out the dependencies
+            var deps = depCollection.Find<BsonDocument>(Builders<BsonDocument>.Filter.In("Properties.BHoM_Guid", tempProject.GetTaskValues()));
+            foreach( BsonDocument doc in deps.ToList())
+                tempProject.AddObject(BHoMObject.FromJSON(doc.ToString(), tempProject));
+            tempProject.RunTasks();
+
+            return ret;
         }
 
-        /*******************************************/
-
-        public List<object> Query(List<string> queryStrings = null, bool keepAsString = false)
+        private BsonDocument ToBson(BHoMObject obj, string key)
         {
-            if (m_Client.Cluster.Description.State == MongoDB.Driver.Core.Clusters.ClusterState.Disconnected)
-                return new List<object>();
-
-            var pipeline = queryStrings.Select(s => BsonDocument.Parse(s)).ToList();
-            List<BsonDocument> result = m_Collection.Aggregate<BsonDocument>(pipeline).ToList();
-            if (keepAsString)
-                return result.Select(x => x.ToString()).ToList<object>();
-            else
-                return result.Select(x => FromBson(x)).ToList<object>();
-        }
-
-        /*******************************************/
-
-        /*public List<string> GetHistoryTimes()
-        {
-            List<object> times = Query(new List<string> { "{$group: {_id: \"$__Time__\"}}", "{$sort: {_id: -1}}" });
-
-        }*/
-
-        /*******************************************/
-        /****  Private Helper Methods           ****/
-        /*******************************************/
-
-        private BsonDocument ToBson(object obj, string key, DateTime timestamp)
-        {
-            if (obj is string) return ToBson(obj as string, key, timestamp);
-            var document = BsonDocument.Parse(BHoM.Base.JSONWriter.Write(obj));  
+            var document = BsonDocument.Parse(obj.ToJSON());
             if (key != "")
-            {
-                document["__Key__"] = key;
-                document["__Time__"] = timestamp;
-            }
-                
-            
+                document["Key"] = key;
             return document;
         }
-
-        /*******************************************/
-
-        private BsonDocument ToBson(string obj, string key, DateTime timestamp)
-        {
-            var document = BsonDocument.Parse(obj);
-            if (key != "")
-            {
-                document["__Key__"] = key;
-                document["__Time__"] = timestamp;
-            }
-                
-            return document;
-        }
-
-        /*******************************************/
-
-        private object FromBson(BsonDocument bson)
-        {
-            MongoDB.Bson.IO.JsonWriterSettings writerSettings = new MongoDB.Bson.IO.JsonWriterSettings { OutputMode = MongoDB.Bson.IO.JsonOutputMode.Strict };
-            return BHoM.Base.JsonReader.ReadObject(bson.ToJson(writerSettings));
-        }
-
     }
 }
