+++ Databases_Engine/Mongo_Adapter.csproj
@@ -7,8 +7,8 @@
     <ProjectGuid>{8CE32CAD-9191-4690-88A2-106245CD6F9F}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Mongo_Adapter</RootNamespace>
-    <AssemblyName>Mongo_Adapter</AssemblyName>
+    <RootNamespace>Databases_Engine</RootNamespace>
+    <AssemblyName>Databases_Engine</AssemblyName>
     <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <TargetFrameworkProfile />
@@ -31,8 +31,7 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
-    <Reference Include="BHoM, Version=1.0.0.0, Culture=neutral, processorArchitecture=MSIL">
-      <SpecificVersion>False</SpecificVersion>
+    <Reference Include="BHoM">
       <HintPath>..\..\BHoM\Build\BHoM.dll</HintPath>
     </Reference>
     <Reference Include="MongoDB.Bson, Version=2.2.4.26, Culture=neutral, processorArchitecture=MSIL">
@@ -57,13 +56,8 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="MongoDataBase.cs" />
-    <Compile Include="MongoLink_Old.cs" />
     <Compile Include="MongoLink.cs" />
-    <Compile Include="MongoServer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Structural\Interface\MongoStructuralElementAdapter.cs" />
-    <Compile Include="Structural\Interface\MongoStructuralResultAdapter.cs" />
   </ItemGroup>
   <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
