+++ Mongo_Toolkit.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 14.0.24720.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mongo_Adapter", "Databases_Engine\Mongo_Adapter.csproj", "{8CE32CAD-9191-4690-88A2-106245CD6F9F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mongo_Test", "Mongo_Test\Mongo_Test.csproj", "{904B7BEB-4B17-461D-840A-CAB29FCA0995}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -17,10 +15,6 @@ Global
 		{8CE32CAD-9191-4690-88A2-106245CD6F9F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8CE32CAD-9191-4690-88A2-106245CD6F9F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8CE32CAD-9191-4690-88A2-106245CD6F9F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{904B7BEB-4B17-461D-840A-CAB29FCA0995}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{904B7BEB-4B17-461D-840A-CAB29FCA0995}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{904B7BEB-4B17-461D-840A-CAB29FCA0995}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{904B7BEB-4B17-461D-840A-CAB29FCA0995}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
